feat: about the app update
